Cargo: move build products to .build/
We move the build products themselves (the ones under srcdir/target/
by default) under the top-level .build/ via .cargo/config.toml.

The Cargo.lock (which we don't commit because we want to see
immediately when we break due to updates) is "moved" via a symlink
into .build/. Note that we must create the .build/…/ directory through
which the symlink points before doing a `cargo build` or it will fail,
since it simply tries to update the file, which will fail if the
leading dirs to the target aren't there. (There really ought to be a
better way to do this, but I've not found one.)

Also add `-e` to the bash options for the build so that if anything in
the embedded mini-script fails, we stop there rather than continuing on.

All this is in preparation to be able to pack up the source and send
it in the Docker context of the build image build so that we can
compile as part of building that image.