Test: Prefix "$USER-" to image and container names
A development machine may have more than one user on it working on
this repo. (This is particularly true where the developers are on Macs
and given a remote server on which to work.) Thus, we must try to
avoid collisions between the images and containers that each developer
is creating at the same time.

We do this by prefixing "$USER-" to both the image and container names.
(We still get layer sharing between users up to the point of creating
the user in the image, which is generally good, but we can turn that
off for a build with the -C option if necessary.)

Another possibility for images was to use $USER as the tag, but then
we have two separate ways of generating the name for images and
containers, which is confusing and tends to provoke errors.