Test: Fix $LINENO of failure in 'trap … ERR' message
In `trap … ERR`, Bash doesn't count as part of $LINENO the initial and
trailing blank lines of our quoted section, or the comment we added to
explain what's going on with the quoting. So we print $LINENO - 4 in
the error message to adjust for this.

We looked at getting rid of the blank lines, but that ends up with
code that's not as clear (the `trap` and `ERR` parts get buried in
with the quoted handling code), so this seems a better solution.