Test: Build and use Docker context under .build/docker/…
Previously we used docker/ as the context, selecting the correct
Dockerfile for the image with the `docker build --file` option.

We now create a separate context under .build/docker/…/ into
which we copy docker/ and then copy the correct Dockerfile.*
to Dockerfile.

This enables us, in future commits, to do additional things to build
the context for each image, such as building the Dockerfile from
several components (some common, some specific to that particular
image), copying in code that should not live under docker/, and so on.