docker/: Build the Dockerfiles from parts so we can share them
The "master" files change from Dockerfile.NAME to dfspec/NAME and now
just include parts from dfparts/. The included files are explicitly
specified so that there's no technical need for the `10-` etc.
prefixes on the dfparts/* files, but demonstrating the intended order
they're to be used is helpful to the developer, since there's a fair
amount of layer ordering optimisation going on here.

Since Dockerfiles don't have any build-in include mechanism, the
preprocessing and including is done with m4. Yes, that's not exactly
the brilliant tool it was in 1977, but the only obvious replacement¹
is a custom templating tool, which likely brings even more complexity
into the system.

Yes, some of these names could be better. Suggestions welcome.

-----
¹ If you find yourself screaming "C preprocessor!" at your screen
right now, you are even more deranged than m4 users and you need to
seek professional help. :-)  Also, that won't work anyway because in
Dockerfiles `#` is a comment character, and changing that would bring
in a world of other misery.