docker/dfparts/60-user-setup: Use bash instead of sh as shell
Turns out that all the RUN lines use /bin/sh by default, but as
we move towards building inside the container we want to be able
to use bash-isms, such as `source` instead of `.` for clarity.

(We didn't notice this before because when we enter the container,
we usually run /bin/bash, and so I'd just sort of assumed that
this was used everwhere. But it's not in Dockerfile RUN lines.)