Test: Further $image_prefix cleanup
docker_build_run() now no longer knows about the $USER part of the
prefix; that's integrated with the `bft` bit of it right at the start
(and now the $USER part is also overridden by the `-i` option). This
makes the handling of this generally cleaner because it's now in one
place.

Also marked $dcontext in docker_build_run() as a local variable.