docker//install: Install .deb package from *-build image
We get the .deb file from the *-build image via a FROM and a
COPY --from. This has the unfortunate side effect of producing
a warning that isn't easy to suppress; details are in the
docker/dfspec/install Dockerfile.

Another option would be to extract the .deb file from the build
image and put it in the context to be uploaded to the install
image build; it's not yet clear (beyond the warning) what
the particular advantages and disadvantages of this might be.

The name of the .deb file (including version number and platform)
is currently hard-coded; thus if we change the version number
of the package the install image build will fail. It's not clear
how we want to fix this.