Test: Remove package/ from webhello.tar.gz source tarball
The Rust build from source is no longer using anything from package/,
and having package/ there was triggering a lot of unnecessary source
rebuilds while working on the packaging.

We now have a separate package.tar.gz that we copy _after_ the source
build so that changes there do not trigger a Rust rebuild.

Co-authored-by: Mike Hume <mhumesf@gmail.com>